feat(auction): 경매를 dashboard 내부(InvestmentTab)로 통합 — 사이드바 유지
문제: 사이드바의 "부동산 경매" 클릭 시 /investment/auction 별도 페이지로 이동해
좌측 사이드바가 사라졌음. 다른 투자 메뉴(전략·계좌·자동매매 등)는 모두
/dashboard?tab=investment&sub=X 통합 페이지에서 동작.

수정
- src/components/Investment/Auction/AuctionContent.tsx 신규: list/detail 통합 컴포넌트
  · selectedItemId 상태로 화면 토글, URL ?item=<uuid> 동기화
- InvestmentTab.tsx: SubTab union/SUB_TABS/SUB_TAB_IDS 에 'auction' 추가, 분기 렌더
- menuConfig.ts: auction route /investment/auction → /dashboard?tab=investment&sub=auction
- investment/layout.tsx: NAV_ITEMS 의 auction href 동일 변경
- AuctionCard: onClick prop 추가 (dashboard 내부 라우팅), 없으면 기존 Link 동작 유지
- /investment/auction, /investment/auction/[itemId]: dashboard 로 redirect (즐겨찾기/외부 링크 호환)

영향: 즐겨찾기 페이지(/investment/auction/favorites)는 그대로 유지 (Link → 상세 진입)
