+++ dental-clinic-manager/src/app/investment/auction/[itemId]/page.tsx
@@ -1,92 +1,8 @@
-'use client'
-import { useEffect, useState, use } from 'react'
-import { useRouter } from 'next/navigation'
-import { ArrowLeft, Loader2 } from 'lucide-react'
-import { AuctionDetailHeader } from '@/components/Investment/Auction/AuctionDetailHeader'
-import { OverviewTab } from '@/components/Investment/Auction/tabs/OverviewTab'
-import { SimulatorTab } from '@/components/Investment/Auction/tabs/SimulatorTab'
-import { RightsTab } from '@/components/Investment/Auction/tabs/RightsTab'
-import { HistoryTab } from '@/components/Investment/Auction/tabs/HistoryTab'
-import { AttachmentsTab } from '@/components/Investment/Auction/tabs/AttachmentsTab'
-import type { AuctionItem, MarketPrice, SimulatorInput } from '@/types/auction'
+// 경매 상세는 dashboard 내부(/dashboard?tab=investment&sub=auction&item=<id>)에서 통합 렌더되므로
+// 이 라우트는 즐겨찾기 / 외부 링크 호환성을 위해 dashboard 로 redirect 한다.
+import { redirect } from 'next/navigation'
 
-interface DetailResponse {
-  item: AuctionItem
-  market: MarketPrice | null
-  rights: any | null
-  history: any[]
-  ai: any | null
-}
-
-const TABS = [
-  { id: 'overview', label: '개요' },
-  { id: 'simulator', label: '시뮬레이터' },
-  { id: 'rights', label: '권리분석' },
-  { id: 'history', label: '이력·통계' },
-  { id: 'attachments', label: '첨부' },
-] as const
-
-type TabId = typeof TABS[number]['id']
-
-export default function AuctionDetailPage({ params }: { params: Promise<{ itemId: string }> }) {
-  const { itemId } = use(params)
-  const router = useRouter()
-  const [data, setData] = useState<DetailResponse | null>(null)
-  const [tab, setTab] = useState<TabId>('overview')
-
-  useEffect(() => {
-    fetch(`/api/auction/items/${itemId}`).then(r => r.json()).then(setData).catch(() => setData(null))
-  }, [itemId])
-
-  if (!data) return <div className="flex justify-center py-16"><Loader2 className="w-6 h-6 animate-spin text-at-accent" /></div>
-
-  return (
-    <div className="max-w-5xl mx-auto px-3 sm:px-0">
-      <button onClick={() => router.back()} className="flex items-center gap-1 text-[14px] md:text-sm font-medium text-at-text-secondary hover:text-at-text mb-3">
-        <ArrowLeft className="w-4 h-4" /> 뒤로
-      </button>
-
-      <AuctionDetailHeader item={data.item} market={data.market} />
-
-      <div className="flex gap-1 border-b border-at-border mb-4 overflow-x-auto -mx-3 sm:mx-0 px-3 sm:px-0">
-        {TABS.map(t => (
-          <button
-            key={t.id}
-            onClick={() => setTab(t.id)}
-            className={`px-4 py-2.5 text-[15px] md:text-sm font-semibold border-b-2 transition-colors whitespace-nowrap shrink-0 ${
-              tab === t.id ? 'border-at-accent text-at-accent' : 'border-transparent text-at-text-secondary hover:text-at-text'
-            }`}
-          >
-            {t.label}
-          </button>
-        ))}
-      </div>
-
-      <div>
-        {tab === 'overview'    && <OverviewTab item={data.item} market={data.market} />}
-        {tab === 'simulator' && (
-          <SimulatorTab
-            item={data.item}
-            market={data.market}
-            onSave={async (input: SimulatorInput) => {
-              await fetch('/api/auction/favorites', {
-                method: 'POST',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({
-                  itemId: data.item.id,
-                  target_bid_price: input.bid_price,
-                  expected_extra_cost: input.repair_cost + input.unpaid_dues,
-                  expected_monthly_rent: input.monthly_rent,
-                }),
-              })
-              alert('관심물건에 저장되었습니다.')
-            }}
-          />
-        )}
-        {tab === 'rights'      && <RightsTab itemId={data.item.id} rights={data.rights} initialAi={data.ai} />}
-        {tab === 'history'     && <HistoryTab itemId={data.item.id} history={data.history} />}
-        {tab === 'attachments' && <AttachmentsTab item={data.item} />}
-      </div>
-    </div>
-  )
+export default async function AuctionDetailRedirect({ params }: { params: Promise<{ itemId: string }> }) {
+  const { itemId } = await params
+  redirect(`/dashboard?tab=investment&sub=auction&item=${encodeURIComponent(itemId)}`)
 }
