+++ dental-clinic-manager/src/app/investment/layout.tsx
@@ -44,7 +44,7 @@ const NAV_ITEMS: NavItem[] = [
   { id: 'rankings', label: '전략 랭킹', icon: Trophy, href: '/dashboard?tab=investment&sub=rankings' },
   { id: 'screener', label: '종목 스크리너', icon: Search, href: '/dashboard?tab=investment&sub=screener' },
   { id: 'smart-money', label: '스마트머니 분석', icon: Brain, href: '/dashboard?tab=investment&sub=smart-money' },
-  { id: 'auction', label: '부동산 경매', icon: Gavel, href: '/investment/auction' },
+  { id: 'auction', label: '부동산 경매', icon: Gavel, href: '/dashboard?tab=investment&sub=auction' },
   { id: 'trading', label: '자동매매', icon: TrendingUp, href: '/dashboard?tab=investment&sub=trading' },
   { id: 'psychology', label: '└ 심리 분석', icon: Users, href: '/dashboard?tab=investment&sub=psychology' },
   { id: 'portfolio', label: '포트폴리오', icon: Briefcase, href: '/dashboard?tab=investment&sub=portfolio' },
